pycryptopp: fix bug in deserialization of verifier key
When Python gives you a char* containing binary data, if you don't cast it to
"unsigned char*" before passing it to the constructor of Verifier, then the
wrong constructor will be matched and the data will be truncated at the first
zero byte, so the data is silently corrupted.  If you do cast it to unsigned
char* then it will still match the wrong constructor, but it will get a type
error saying that you can't automatically convert unsigned char* to to char*.
Note, however, that it silently converted the second argument, which is the
size of the data, to bool ("pumpAll").

So to get the right constructor, you need to cast to unsigned char* and pass a
third argument ("pumpAll") so that it matches the first argument to the data
and the second argument to the length.

I think there's a lesson for software engineers in here, but I'm not sure what
it is.