pycryptopp: remove the "foo.py" empty place-holder modules
We formerly had modules like "aes.py" and "sha256.py", which did nothing except
"from _aes import *".  This was because the _aes module was produced from C++
code, and we wanted people to find what they were looking for if they ran 
"find . -name '*.py'".  However, this is not a very important feature, and it
turns out that it makes the result of "help(modulename)" quite ugly.

Now the C++ code generates modules named "aes" and "sha256" and so on.