setup: new option --disable-embedded-cryptopp: use include files and libcryptopp from the system instead of using the Crypto++ source code which is bundled in the pycryptopp source tree.
This is a requirement of repackagers such as Debian and Fedora.

Unfortunately, when I try to use --disable-embedded-cryptopp on my Ubuntu Hardy machine, I run 
afoul of this problem:

http://www.gnu.org/software/gcc/faq.html#dso 

I haven't been able to figure out how to build a sha256module.so, which is linked against 
libcryptopp.so, such that they recognize Crypto++ symbols as being "the same" when referenced in 
libcryptopp.so and in sha256module.so.