add licence options: MIT or SPL for the whole package
The choice of MIT is because Twisted Python uses it, and folks working on the
Twisted project have expressed interest in using pycryptopp:

http://twistedmatrix.com/trac/ticket/4633

The choice of SPL is because it is so darn simple.

Also a few minor tweaks to docs and setup.py: fix some URLs that had typos in
them, utf-8 decode the README before passing it to setup() and hence to the
Python Package Index, remove the Trove Classifier indicating that we support
Python 2.4 and Python 2.5. Reformat TGPPL from html to rst.